+++ site/public/ts/utils/actions/addition/ConnectionAction.ts
@@ -1,8 +1,8 @@
-import {Action} from "./Action";
-import {CircuitDesigner} from "../../models/CircuitDesigner";
-import {InputPort} from "../../models/ports/InputPort";
-import {OutputPort} from "../../models/ports/OutputPort";
-import {Wire} from "../../models/ioobjects/Wire";
+import {Action} from "../Action";
+import {CircuitDesigner} from "../../../models/CircuitDesigner";
+import {InputPort} from "../../../models/ports/InputPort";
+import {OutputPort} from "../../../models/ports/OutputPort";
+import {Wire} from "../../../models/ioobjects/Wire";
 
 export class ConnectionAction implements Action {
     private designer: CircuitDesigner;
