+++ site/public/ts/utils/tools/SplitWireTool.ts
@@ -1,7 +1,5 @@
 import {DEFAULT_SIZE} from "../Constants";
 
-import {BezierContains} from "../math/MathUtils";
-
 import {Tool} from "./Tool";
 import {TranslateTool} from "./TranslateTool"
 import {SelectionTool} from "./SelectionTool";
@@ -16,7 +14,7 @@ import {WirePort} from "../../models/ioobjects/other/WirePort";
 import {Action} from "../actions/Action";
 import {GroupAction} from "../actions/GroupAction";
 import {SelectAction} from "../actions/SelectAction";
-import {SplitWireAction} from "../actions/SplitWireAction";
+import {SplitWireAction} from "../actions/addition/SplitWireAction";
 
 export class SplitWireTool extends TranslateTool {
     private splitAction: GroupAction;
