+++ tests/public/ts/utils/actions/PlaceAction.test.ts
@@ -2,7 +2,7 @@ import "jest";
 
 import {CircuitDesigner} from "../../../../../site/public/ts/models/CircuitDesigner";
 import {Switch}          from "../../../../../site/public/ts/models/ioobjects/inputs/Switch";
-import {PlaceAction}     from "../../../../../site/public/ts/utils/actions/PlaceAction";
+import {PlaceAction}     from "../../../../../site/public/ts/utils/actions/addition/PlaceAction";
 
 describe("PlaceAction", () => {
     it("Undo/Redo", () => {
